chore: introduce separate api crates for control_panel, upgrader, and wallet canisters (#37)
This PR introduces separate api crates (containing what used to be the
`transport` module) for control panel, upgrader, and the wallet
canister. The motivation behind this change is to enable integration
tests (to be bootstrapped in a follow up MR) depend on the api types of
the canisters without depending on their entire implementation (which
makes the tests not build for non-WASM targets). For instance, OpenChat
has the same code structure keeping the api types separate from the
actual implementation of the canister logic.