feat: add account proposal operations (#44)
This PR adds the edit account operation type and changes the creation of
an account to use the `AddAccount` proposal.

Also updated the frontend to use those proposals. No policies currently
apply to them since the policy design is still in progress. I've added
the todo's in the code and will come back to it later once the policy
system is finalized.

---------

Co-authored-by: mraszyk <31483726+mraszyk@users.noreply.github.com>
Co-authored-by: Martin Raszyk <martin.raszyk@dfinity.org>