feat: expired session overlay (#104)
Display overlay when the session expires.

Internet Identity sessions have a fixed timespan. Upon expiration the
user needs to log in again to continue using the dapp. To not break the
workflow of the user the dapp displays a persistent overlay asking to
reauthenticate. After reauthentication the user can continue as normal.

How it works:
* on initialization a timer is set to fire when the internet identity
session expires
* using a BroadcastChannel the frontend notifies other open tabs when
the user signs in, signs out, or performs any activity
* when the session expiration timer expires a persistent dialog is
displayed with a button to reauthenticate
* after reauthentication the session expiration timer is reset and other
tabs are notified
* when there is no activity in the current tab or on other tabs for a
certain time, the session gets cleared and the dialog is shown

Todo
- [x] handle sign out action across browser tabs
- [x] expire session on inactivity
- [x] if the user reauthenticates with a new principal, reset the app
before login
- [x] tests
- [x] remove verbose logging

---------

Co-authored-by: Kepler Vital <dev@keplervital.com>