feat: add description field to wallet (#151)
Enable users to add an optional summary to transfer requests. Summary
data is stored in Proposal::summary.

Updated create transfer form, it is also prepared to load and show an
existing transfer along with the summary from its proposal. To support
this I added the proposal_id to the Transfer DTO.

![image](https://github.com/dfinity/orbit-wallet/assets/9403182/fafa7ec4-6ff5-432e-88d0-dba340a9d939)

Display the summary on the Proposal details dialog:

![image](https://github.com/dfinity/orbit-wallet/assets/9403182/f0094810-1599-4613-9f30-8ddf295dd77d)