fix: canisters with specified IDs must have canister IDs from mainnet ranges (#177)
This PR makes canister IDs of canisters created with specified IDs (flag
`--specified-id`) belong to mainnet canister ranges (at the end of the
canister range of the subnet `snjp4` so that they also don't collide
with actually used canister IDs on mainnet that are allocated from the
beginning of the corresponding canister range). This is necessary since
otherwise canister creation without specifying a canister ID (when not
using the `--specified-id` feature) may fail in local testing.