chore: rename id to identity in user struct of control panel (#192)
This PR renames the field `id` to `identity` in the `User` struct of the
control panel canister. This way, a follow up PR could introduce the
field `id: UUID` into the `User` struct of the control panel canister
and a further follow up PR could use this UUID in user subscription
status indices.