feat: no upgrader upgrades triggered by wallet upgrade (#204)
This MR removes the possibility to upgrade the upgrader canister when
upgrading the wallet (note that the upgrader canister can be upgraded by
a wallet proposal at any time). The motivation for this change is to
simplify the steps performed when initializing the canister after an
upgrade and thus making the initialization more robust.

This MR also declares the method `register_jobs` as not `async` since it
is already not `async`.