feat: add verified actor option to canister calls (#215)
We've have been using many calls as query calls to the canister,
however, this is not secure in many situations. To address this issue
this PR creates a dynamic strategy for when to use update calls, and
starts of lists using query calls but then on the background it get's
refresh with update calls and dialogs where the user would change
information it fetches the data with update calls.