feat: add user station labels (#258)
This PR adds the possibility to add labels to the associated stations of
the user in the control panel. This enables features such as storing
user preferences for the given station.

Also, the concept of main station was removed, the main station is now
the first in the list and the list is ordered.