fix: canbench station build (#262)
This PR fixes the station canister build for canbench since canbench now
(v0.1.4) imposes the system subnet's limits on the canister WASM size
(3.5M). Independently of that, it is a good idea to use the same script
for benchmarking and production build, only adding the canbech feature
in the benchmarking build. It is also a good idea to pin the canbench
version so that the CI is not flaky.