feat(wallet): add versioned path to wallet ui (#263)
Since station canisters that the wallet UI connects with can be at many
different versions, even forward compatible changes make it difficult
since the UI could expect a newer API then what it got.

This PR addresses that by introducing versioned UIs that keep track of
the api version they are compatible with. It adds a compatibility layer
that upon connecting to a different station canister, fetch's it's
version and validates what's the best UI to use with it.

The `compatibility.core` is where most of the logic happens, however,
i've also had to refactor the vite build to be able to operate
seamlessly with the many versioned paths.