feat(control-panel): add registry api interface and repository (#277)
This PR adds the registry api interface and repositories for the
registry. I've only included some of the get implementation in the
service layer for now and had to move the pagination from the station
canister to the common orbit_essentials lib to be reused by the
control-panel.

Also, i'll still create a follow up PR that will remove the pagination
implementation from the station canister and will start to use it from
the common orbit_essentials lib.