chore(station): introduce ValidationError type (#280)
This PR introduces a new type `ValidationError` as a union type of the
existing `RecordValidationError` type and a new type
`ExternalCanisterValidationError` for invalid external canisters (to be
used in a follow-up PR introducing call requests to external canisters).

Note that adding a new variant for invalid external canisters to the
existing `RecordValidationError` would require converting
`RecordValidationError::NotFound` to `ExternalCanisterError` and also
`ExternalCanisterValidationError::InvalidExternalCanister` to
`TransferError` which makes little sense. Hence, it makes more sense to
introduce a new type `ValidationError` as a union type and distinguish
between `RecordValidationError` and `ExternalCanisterValidationError` on
the type level.