feat(station): disaster recovery MVP (#271)
This PR implements minimal disaster recovery functions through syncing
admin users and accounts into the upgrader, and allowing the admins to
agree on recovering the station canister and its accounts by submitting
recovery requests directly to the upgrader.

Changes:
- new SetDisasterRecovery request type which marks a user group as the
Disaster Recovery Committe, and the quorum_percentage required for
triggering the recovery process
- on insert/remove of users, sync the users that belong to the committee
to the upgrader
- on insert of an account, sync the accounts to the upgrader so that
they can be recovered in the event of losing the station canister state
- committee members can use their identities associated with their user
to submit disaster recovery requests to the Upgrader. When quorum is
reached, the upgrader will perform the Install, Upgrade, or Reinstall of
the station canister
- in order to reach consensus (quorum) committee members must submit the
same wasm module, args, and install type in their request
- new logging infrastructure for the Upgrader that can be queried by
committee members
- updated project structure following the structure of the station and
control panel
- the user group which is marked as the committee user group cannot be
removed
- syncs all after the upgrader is upgraded, in case the upgrader didn't
previously support disaster recovery
- creates default policies based on the quorum specified by
`quorum_percentage` and the number of admins

Todo:

- [x] check backwards compatibility

---------

Co-authored-by: Kepler Vital <dev@keplervital.com>