feat(station): notify failed station upgrade (#331)
This PR makes the upgrader canister notify the station canister about a
failed station upgrade. Consequently, the station can mark the (so far
processing) station upgrade request as failed.

Note that the `trigger_upgrade` method of the upgrader needs to perform
the upgrade _after_ responding to the station so that the station can be
stopped. But this means that the response of the upgrader to the station
cannot contain any subsequent failures encountered during the actual
upgrade. Hence, such failures need to be communicated in a separate call
from the upgrader to the station.