feat(dfx-orbit): dfx-orbitv0.3 (#318)
This PR merges a new version of dfx-orbit.

Improvements:
- Proper logging, with settable verbosity (`v` and `q` flags). Logging
goes to StdErr, output to StdOut
- Pass `--station [Name]` to address a station that is not the default
station
- A `review` commands no longer by default dump the json to output, but
rather provide a human readable Gist. The information is annotated with
information, such as candid is parsed, and canister ids are reverse
looked up to annotate the ids with the names of the canisters.