chore(station): use arg length as request size in rate limiter (#363)
This PR replaces a custom request "size" (which would have to be refined
to cover all parts of a request and updated every time we make a change
to the request type) by the update call arg length for the sake of the
request size rate limiter. This way, all parts of a request (incl. any
future extensions) are covered automatically.