chore(dfx-orbit): dfx-orbit version 0.6.0 (#377)
This version of dfx-orbit is functionally equivalent with version 0.5.0.

It just has the code completely refactored to have code that belongs to
the same feature live together, rather than code that has the function,
i.e. `request canister call` and `verify canister call` are in the same
submodule, rather than all `request ...` and all `verify ...`
functionality.