chore(dfx-orbit): Update dependencies to use SDK master with repro fix (#400)
This PR bumps ic-agent, ic-utils, and dfinity/sdk dependencies to their
more recent versions to fix a bug in dfx-orbit: chunks were not uploaded
to the asset canister in the right order. The fix has been tested
manually.

Also it means that we are using dfinity/sdk master branch with the repro
fix now.