fix(station): Make DR sync + logging more efficient (#432)
Problem:

1. The station syncs all assets and accounts to the upgrader on any
account and asset change. The accounts change on each balance query
since the new balance is saved back to the repository. This results in
very frequent unnecessary syncs.
2. The upgrader log size is growing rapidly. Trimming is necessary, but
it cannot be trimmed due to the stable Log structure only supporting
append operation.

Task:
1. [x] only sync assets and accounts when the change is relevant - ie.
any field of the DR version of the entries change
2. [x] switch to a stable BTreeMap to keep the logs and limit the log
size to 25k items

Also bundled a small fix for a missing asset foreign key check.

---------

Co-authored-by: Kepler Vital <dev@keplervital.com>