fix(wallet): fix request export order (#449)
- implements the sort_by optional field for
GetNextApprovableRequestInput to be able to affect the order it will
produce the next approvable request
- removes the extra sorting on the frontend for csv export generation
that will put the results in reverse order
- bug fix: `list_requests` with `only_approvable = true` could return
non-pending requests for users who could approve if the request was
still pending


Manual testing findings:

1. Uploaded CSV file for batch transfers. The transfers are created from
the frontend 10 in parallel at a time. The creation order inside a batch
of 10 is random. The batch transfer feature therefore will create a list
of transfers that dont keep the ordering of the original CSV input file.
2. Exporting the created transfers will produce a list that is sorted
from most recent to oldest. This is in line with the ordering of the
data table on the frontend.
3. Using the load-next button will enumerate the pending requests in the
same order as in the exported CSV file. However if you start approving
from the middle of the list, the next approvable request will jump back
to the first one in the list.