chore(station,upgrader): do not sync station data to upgrader after every upgrade of upgrader (#550)
This PR stops syncing station data to the upgrader after every upgrade
of the upgrader. This is safe under the assumption that people don't use
a recent station (containing this PR) to upgrade an old upgrader that
did not contain any station data and thus syncing station data after an
upgrade was necessary.