fix(station): harden async batch flows (#572)
This PR hardens async batch flows to not enumerate the results and index
into foreign vectors since those foreign vectors might not be aligned
with the result vector (it does not seem to be an issue at the moment,
but the current design is error-prone w.r.t. future changes).